int64_t since genome size cannot fit into int variable

+++ programs/submodules/chunk/chunk.c
@@ -950,13 +950,13 @@ int ChunksCreator_getTotalNumberOfChunks(ChunksCreator *chunksCreator) {
     return (int) stList_length(chunksCreator->chunks);
 }
 
-int ChunksCreator_getTotalLength(ChunksCreator *chunksCreator) {
+int64_t ChunksCreator_getTotalLength(ChunksCreator *chunksCreator) {
     ChunkIterator *iterator = ChunkIterator_construct(chunksCreator);
 
     ptBlock *block = NULL;
 
     char ctg[200];
-    int totalLength = 0;
+    int64_t totalLength = 0;
     while ((block = ChunkIterator_getNextPtBlock(iterator, ctg)) != NULL) {
         int start = block->rfs;
         int end = block->rfe;
